# Optimize class_attribute instance accessors