### Improvement

Here's the benchmark I used: https://gist.github.com/casperisfine/1f37a7228f9133ba30441ff796fe0c1b

```
==== definition ====
 old     11.341k (±24.3%) i/s -     52.415k in   5.032040s
 new      1.708k (±18.9%) i/s -      8.300k in   5.058740s

==== class reader ====
 old     14.143M (± 1.7%) i/s -     70.951M in   5.018039s
 new     14.126M (± 1.4%) i/s -     70.956M in   5.024121s

==== class inherited reader ====
 old     14.196M (± 1.3%) i/s -     71.104M in   5.009397s
 new     13.852M (± 1.4%) i/s -     69.525M in   5.020324s

==== class writer ====
 old    721.934k (±22.4%) i/s -      3.375M in   5.074818s
 new    976.412k (±20.8%) i/s -      4.706M in   5.043937s

==== instance reader ====
 old      5.153M (± 0.9%) i/s -     25.763M in   5.000218s
 new      9.962M (± 1.5%) i/s -     50.054M in   5.025634s

==== instance writer ====
 old      9.411M (± 1.6%) i/s -     47.132M in   5.009605s
 new     21.748M (± 1.5%) i/s -    108.722M in   5.000272s
```

In short, the instance accessors are twice faster with this patch, and the class writer is marginally faster.

The downside is that the definition of the class variable is much slower. So this trade a bit of boot time performance for runtime performance. But it's not like the definition of `class_attribute` is really a hotspot.

### This implementation

Most of the gains are thanks to `class_eval` that allows to replace `instance_variable_defined?(:@foo)` by `defined?(@foo)`, `instance_variable_set(:@foo, value)` by `attr_writer`, and similar changes.

### Discarded implementations

Initially I was actually trying to reclaim some performance on the class accessors. I found an implementation based on `defined?(@foo) ? @foo : superclass.foo`, that was twice faster on the class itself, but then would slowly degrade when calling on child classes.

cc @rafaelfranca @Edouard-chin @etiennebarrie @paracycle @Morriar