@rafaelfranca 2019-12-06
@tenderlove do you have any idea how we can improve the class accessor?