@casperisfine 2019-12-06
Uh weird failures on CI, I didn't expect that since the ActiveSupport test suite was green. I'll take a look.