@casperisfine 2019-12-16
I benched the class accessor against a regular `class << self; attr_reader; end` and it's in the same ballpark:

```
attr_reader     21.296M (± 1.9%) i/s -    106.875M in   
class_attribute reader     14.145M (± 1.7%) i/s -     70.880M in   5.012404s
```

So there's probably no gain to be found on the class reader.

I rebased the PR and fixed the syntax error.