@kaspth 2019-12-22
@casperisfine thanks! There were some stylistic things I felt strongly about, so I committed this in https://github.com/rails/rails/commit/49d1b5a98dc50ca59bbfb817f90eb5c2c2645d30

(Accidentally lost the merge commit somehow)